add raw contract test
